refactor: record service interface 생성
