+++ src/main/java/corecord/dev/domain/folder/application/FolderServiceImpl.java
@@ -13,14 +13,12 @@ import corecord.dev.domain.record.application.RecordDbService;
 import corecord.dev.domain.user.application.UserDbService;
 import corecord.dev.domain.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
-@Slf4j
 @RequiredArgsConstructor
 public class FolderServiceImpl implements FolderService {
